requirements.txt
audiofile==1.1.1
audmetric==1.1.4
audobject==0.7.7
audtorch==0.6.4
hear21passt==0.0.16
librosa==0.9.2
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2
PyYAML==6.0
scikit-learn==1.2.0
scipy==1.9.3
timm~=0.4.12
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchinfo==1.7.1
torchlibrosa==0.0.9
torchvision==0.14.1+cu117
torchviz==0.0.2
tqdm==4.64.1
wget==3.2
seaborn~=0.12.2
fairseq~=0.12.2
audplot~=1.4.6
torchsummary~=1.5.1
omegaconf~=2.0.6